+++ js/utils/tpl.js
@@ -1,5 +1,5 @@
-var TemplateEngine = function(html, options) {
-    var re = /<%(.+?)%>/g,
+var TemplateEngine = function(html, options, conf) {
+    var ReEx, re = '<%(.+?)%>',
         reExp = /(^( )?(var|if|for|else|switch|case|break|{|}|;))(.*)?/g,
         code = 'with(obj) { var r=[];\n',
         cursor = 0,
@@ -10,9 +10,23 @@ var TemplateEngine = function(html, options) {
             (code += line !== '' ? 'r.push("' + line.replace(/"/g, '\\"') + '");\n' : '');
         return add;
     };
-    while(match = re.exec(html)) {
+
+    if (Utils.isValue(conf)) {
+        if ((conf.hasOwnProperty('beginToken'))) {
+            re = re.replace('<%', conf.beginToken);
+        }
+        if ((conf.hasOwnProperty('endToken'))) {
+            re = re.replace('%>', conf.endToken);
+        }
+    }
+
+    ReEx = new RegExp(re, 'g');
+    match = ReEx.exec(html);
+
+    while(match) {
         add(html.slice(cursor, match.index))(match[1], true);
         cursor = match.index + match[0].length;
+        match = ReEx.exec(html);
     }
     add(html.substr(cursor, html.length - cursor));
     code = (code + 'return r.join(""); }').replace(/[\r\t\n]/g, ' ');